testingclient: Move matching key types and values to keys.go
These will be used by multiple clients.

Also add remaining verbs used by client-go testing.Action implementations.

Authored-by: David Sharp <dsharp@vmware.com>
